refactor: validate 분리
